WIP: refactored 'wrapLink', 'unwrapLink', 'hasLinks', 'onPaste'
