+++ components/SlateEditor.jsx
@@ -1,8 +1,7 @@
 import React from "react";
-import { Editor, getEventTransfer } from "slate-react";
+import { Editor } from "slate-react";
 import { Value } from "slate";
 import { isKeyHotkey } from "is-hotkey";
-import isUrl from "is-url";
 import PropTypes from "prop-types";
 
 import { initialValue, Button, Toolbar } from "./slate-editor-core";
@@ -10,34 +9,11 @@ import { initialValue, Button, Toolbar } from "./slate-editor-core";
 // TextAlign
 import { hasMultipleAligns, renderAlignButton, renderAlignButtons } from "./slate-editor-formats/text-align";
 
-/**
- * A change helper to standardize wrapping links.
- *
- * @param {Editor} editor
- * @param {String} href
- */
-
-function wrapLink(editor, href) {
-	editor.wrapInline({
-		type: "link",
-		data: { href },
-	});
-
-	editor.moveToEnd();
-}
-
-/**
- * A change helper to standardize unwrapping links.
- *
- * @param {Editor} editor
- */
-
-function unwrapLink(editor) {
-	editor.unwrapInline("link");
-}
-
-
-// The rich text example.
+// Link
+import wrapLink from "./slate-editor-formats/link/wrapLink";
+import unwrapLink from "./slate-editor-formats/link/unwrapLink";
+import hasLinks from "./slate-editor-formats/link/hasLinks";
+import onPaste from "./slate-editor-formats/link/onPaste";
 
 class SlateEditor extends React.Component {
 	// Constructor
@@ -68,35 +44,6 @@ class SlateEditor extends React.Component {
 	onChange = ({ value }) => {
 		this.setState({ value });
 	};
-
-	/**
-	 * On paste, if the text is a link, wrap the selection in a link.
-	 *
-	 * @param {Event} event
-	 * @param {Editor} editor
-	 * @param {Function} next
-	 */
-
-	hasLinks = () => {
-		const { value } = this.state;
-		return value.inlines.some(inline => inline.type === "link");
-	}
-
-	onPaste = (event, editor, next) => {
-		
-		if (editor.value.selection.isCollapsed) return next();
-
-		const transfer = getEventTransfer(event);
-		const { type, text } = transfer;
-		if (type !== "text" && type !== "html") return next();
-		if (!isUrl(text)) return next();
-
-		if (this.hasLinks()) {
-			editor.command(unwrapLink);
-		}
-		editor.command(wrapLink, text);
-		event.preventDefault();
-	}
 
 	// On key down, if it's a formatting command toggle a mark.
 	onKeyDown = (event, editor, next) => {
@@ -183,10 +130,9 @@ class SlateEditor extends React.Component {
 
 	removeUrl = () => {
 		const { editor } = this;
-		if (this.hasLinks()) {
-			this.unwrapLink(editor);
+		if (hasLinks(this.state.value)) {
+			unwrapLink(editor);
 		}
-		
 	}
 
 	editLink = () => {
@@ -195,7 +141,7 @@ class SlateEditor extends React.Component {
 		const { value } = this.state;
 		let currentHref = "";
 
-		if (this.hasLinks()) {
+		if (hasLinks(this.state.value)) {
 			value.inlines.some(inline => {
 				currentHref = inline.data.get("href");
 			});
@@ -255,9 +201,8 @@ class SlateEditor extends React.Component {
 
 		const { editor } = this;
 		const { value } = editor;
-		const hasLinks = this.hasLinks();
 
-		if (hasLinks) {
+		if (hasLinks(this.state.value)) {
 			editor.command(unwrapLink);
 		} else if (value.selection.isExpanded) {
 			const href = window.prompt("Enter the URL of the link:");
@@ -339,7 +284,7 @@ class SlateEditor extends React.Component {
 	};
 
 	renderInlineButton = icon => {
-		const isActive = this.hasLinks();
+		const isActive = hasLinks(this.state.value);
 		return (
 			<Button
 				active={isActive}
@@ -541,7 +486,7 @@ class SlateEditor extends React.Component {
 						value={this.state.value}
 						onChange={this.onChange}
 						onKeyDown={this.onKeyDown}
-						onPaste={this.onPaste}
+						onPaste={event => onPaste(event, this.editor, this.state.value)}
 						renderBlock={this.renderBlock}
 						renderMark={this.renderMark}
 						renderInline={this.renderInline}
