+++ components/slate-editor-formats/link/unwrapLink.js
@@ -1,16 +1,14 @@
-/**
- *
- * A change helper to standardize unwrapping links.
- *
- * @param {Editor} editor - the global editor reference.
- *
- */
-
 import PropTypes from "prop-types";
 
-const unwrapLink = editor => {
+/**
+ * A change helper to standardize unwrapping links.
+ *
+ * @param {Editor} editor - A global editor reference, e.g. "this.editor".
+ */
+
+function unwrapLink(editor) {
 	editor.unwrapInline("link");
-};
+}
 
 unwrapLink.propTypes = {
 	editor: PropTypes.object.isRequired,
